requirements.txt
django-taggit==1.3.0
Markdown==3.3.3
celery==5.0.4
gevent==20.9.0
redis==3.5.3
Pillow==7.1.2
django-imagekit==4.0.2 
pilkit==2.0